Overlays are now managed by VectorManipulatorShapes
